requirements.txt
azure-ai-evaluation
promptflow-azure
azure-identity
wikipedia
python-dotenv
azure-ai-projects
azure-ai-ml